ai_service/requirements.txt
flask
flask-cors
google-generativeai
numpy
pandas
scikit-learn
python-dotenv
requests
